# docs/api-openapi.yaml
openapi: 3.0.3
info:
  title: PR4Y API
  version: "1.0.0"
servers:
  - url: http://localhost:4000/v1
  - url: https://your-api.railway.app/v1
paths:
  /health:
    get:
      summary: Healthcheck
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  version: { type: string }

  /auth/register:
    post:
      summary: Register with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8, maxLength: 256 }
      responses:
        "200":
          description: Registered; returns tokens and user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already registered
        "400":
          description: Validation error

  /auth/login:
    post:
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
        "400":
          description: Validation error

  /auth/refresh:
    post:
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: New tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid or expired refresh token

  /auth/logout:
    post:
      summary: Logout (revoke refresh token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, enum: [true] }

  /crypto/wrapped-dek:
    get:
      summary: Get wrapped DEK
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Wrapped DEK
          content:
            application/json:
              schema:
                type: object
                required: [kdf, wrappedDekB64]
                properties:
                  kdf:
                    type: object
                    properties:
                      name: { type: string }
                      params: { type: object }
                      saltB64: { type: string }
                  wrappedDekB64: { type: string }
        "404":
          description: Wrapped DEK not found
    put:
      summary: Store wrapped DEK
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kdf, wrappedDekB64]
              properties:
                kdf:
                  type: object
                  required: [name, params, saltB64]
                  properties:
                    name: { type: string }
                    params: { type: object }
                    saltB64: { type: string }
                wrappedDekB64: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }

  /sync/pull:
    get:
      summary: Pull encrypted records
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: cursor
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 100, maximum: 500 }
      responses:
        "200":
          description: Records
          content:
            application/json:
              schema:
                type: object
                properties:
                  nextCursor: { type: string }
                  records:
                    type: array
                    items:
                      $ref: "#/components/schemas/EncryptedRecord"

  /sync/push:
    post:
      summary: Push encrypted records batch
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [records]
              properties:
                records:
                  type: array
                  maxItems: 100
                  items:
                    $ref: "#/components/schemas/EncryptedRecordInput"
      responses:
        "200":
          description: Push result
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted:
                    type: array
                    items: { type: string }
                  rejected:
                    type: array
                    items:
                      type: object
                      properties:
                        recordId: { type: string }
                        reason: { type: string }
                  serverTime: { type: string }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthResponse:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresIn: { type: integer }
        user:
          type: object
          properties:
            id: { type: string }
            email: { type: string }
            createdAt: { type: string }
    EncryptedRecord:
      type: object
      properties:
        recordId: { type: string }
        type: { type: string }
        version: { type: integer }
        encryptedPayloadB64: { type: string }
        clientUpdatedAt: { type: string }
        serverUpdatedAt: { type: string }
        deleted: { type: boolean }
    EncryptedRecordInput:
      type: object
      required: [recordId, type, version, encryptedPayloadB64, clientUpdatedAt, deleted]
      properties:
        recordId: { type: string }
        type: { type: string }
        version: { type: integer }
        encryptedPayloadB64: { type: string }
        clientUpdatedAt: { type: string }
        deleted: { type: boolean }